document-parser-v0/requirements.txt
docling==2.9.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
pydantic==2.5.0
python-dotenv==1.0.0
structlog==24.1.0
click==8.1.7